tier-1/backend/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
pact-python==2.2.0
pytest==8.2.0